backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
anthropic
numpy
python-multipart
PyJWT
bcrypt
twilio
httpx
